Add lock for OPKG.
OPKG commands are now put on a FIFO queue. Before processing an item, we check if the opkg lock file is present, if so we wait a bit re-checking the file. Should the lock not resolve after a given amount of time, we stop processing the queue, rejecting all items.

Processing an opkg command is also guarded by a timeout, should a given command not finish processing withing 60 seconds, we consider the operation failed and reject all other queued OPKG commands. This rejection does not affect the OPKG command running on the device, which might finish even after rejection in the client.

All of this should improve potential race conditions that might arise due to checking for updated when the application starts up and the user doing other package related operations in the mean time, like for example loading the package list.