Add lock for OPKG. (#333)
OPKG commands are now put on a FIFO queue. Before processing an item, we check if the opkg lock file is present, if so we wait a bit re-checking the file. Should the lock not resolve after a given amount of time, we stop processing the queue, rejecting all items.

All of this should improve potential race conditions that might arise due to checking for updated when the application starts up and the user doing other package related operations in the mean time, like for example loading the package list.

* Show error when fetching packages fails
* Show error when fetching package details fails
* Handle rejections during package removal on package detail page
* Handle rejections during package installation on package detail page. Refactored package management errors to their own component
* Use general package management error on package list.
* Add rejection reason to be shown when lock is present. Re-use generic package management error on upgrade.
* Display loading spinner on package list page as soon as page starts loading.